# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "StageUpdateRequest": {
        "description": "推进投递阶段的请求体（`stage` 为 `ApplicationStage` 的字面值）。",
        "properties": {
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "stage"
        ],
        "title": "StageUpdateRequest",
        "type": "object"
      },
      "StageUpdateResponse": {
        "description": "推进投递阶段的结果。",
        "properties": {
          "application_id": {
            "title": "Application Id",
            "type": "string"
          },
          "stage": {
            "title": "Stage",
            "type": "string"
          }
        },
        "required": [
          "application_id",
          "stage"
        ],
        "title": "StageUpdateResponse",
        "type": "object"
      },
...
      "StreamRequest": {
        "description": "对话请求体（流式与一次性共用）。",
        "properties": {
          "conversation_id": {
            "default": "",
            "title": "Conversation Id",
            "type": "string"
          },
          "message": {
            "minLength": 1,
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "StreamRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Hunter1",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/applications/{application_id}/stage": {
      "post": {
        "operationId": "update_stage_api_applications__application_id__stage_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StageUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StageUpdateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "推进投递阶段"
      }
    },
...
    "/api/assistant/stream": {
      "post": {
        "operationId": "stream_api_assistant_stream_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "流式对话（SSE）"
      }
    },
...
    "/api/jobs/{job_id}/apply": {
      "post": {
        "operationId": "job_apply_api_jobs__job_id__apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "记录投递"
      }
    },
...
    "/api/settings": {
      "get": {
        "operationId": "read_settings_api_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/SettingsView"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Read Settings Api Settings Get"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "读取 LLM 配置"
      },
      "put": {
        "operationId": "save_settings_api_settings_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsForm"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "保存 LLM 配置"
      }
    },
...
    "/applications": {
      "get": {
        "operationId": "applications_page_applications_get",
        "parameters": [
          {
            "in": "query",
            "name": "recorded",
            "required": false,
            "schema": {
              "default": "",
              "title": "Recorded",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Applications Page"
      }
    },
    "/applications/{application_id}/delete": {
      "post": {
        "operationId": "application_delete_applications__application_id__delete_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Application Delete"
      }
    },
    "/applications/{application_id}/stage": {
      "post": {
        "operationId": "application_stage_applications__application_id__stage_post",
        "parameters": [
          {
            "in": "path",
            "name": "application_id",
            "required": true,
            "schema": {
              "title": "Application Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_application_stage_applications__application_id__stage_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Application Stage"
      }
    },
    "/assistant": {
      "get": {
        "operationId": "assistant_page_assistant_get",
        "parameters": [
          {
            "in": "query",
            "name": "cid",
            "required": false,
            "schema": {
              "default": "",
              "title": "Cid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "default": "",
              "title": "Error",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assistant Page"
      },
      "post": {
        "operationId": "assistant_send_assistant_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_assistant_send_assistant_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assistant Send"
      }
    },
    "/assistant/stream": {
      "post": {
        "description": "流式回合：以 `text/event-stream` 边生成边推。\n\n与 `/assistant` 的差别只在交付方式 —— 内容逐段到达，工具执行前后各有\n一个事件。**失败也走事件流**（一条 `error` 事件）：响应一旦开始就没法\n再重定向，这是流式端点最容易漏掉的一条。",
        "operationId": "assistant_stream_assistant_stream_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_assistant_stream_assistant_stream_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assistant Stream"
      }
    },
    "/crawl": {
      "get": {
        "operationId": "crawl_page_crawl_get",
        "parameters": [
          {
            "in": "query",
            "name": "started",
            "required": false,
            "schema": {
              "default": "",
              "title": "Started",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "refused",
            "required": false,
            "schema": {
              "default": "",
              "title": "Refused",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Crawl Page"
      },
      "post": {
        "operationId": "crawl_start_crawl_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Crawl Start"
      }
    },
...
  }
}